fix(imports): Fix import issues after potentials module refactor
- Updated import path for `Potential` class in `src/thermoelasticsim/md/integrators.py`.
- Updated import path for `Potential` class in `src/thermoelasticsim/utils/optimizers.py`.
- Ensured all tests pass, verifying the correctness of import fixes.
