+++ src/thermoelasticsim/md/integrators.py
@@ -7,7 +7,7 @@ import numpy as np
 from typing import Optional, List
 import logging
 from thermoelasticsim.core.structure import Cell
-from thermoelasticsim.potentials.potentials import Potential
+from thermoelasticsim.potentials import Potential
 
 logger = logging.getLogger(__name__)
